Create README.md
It is used to apply many transformations to your text in a way to make it usable according to your requirement. It is used to apply many common alterations to text like: converting to UPPERCASE, converting to LOWERCASE, clear TEXT and handle SPACES.
The core feature of this application is its component based structure which makes it efficient and reusable for the developer. It is free to use browser compatible application where you can analyze your text with certain functionalities resulting in an efficient analyzation of you desired text.
HIGHLIGHTING FEATURES:
•	Changing the text to uppercase
•	Counting words and characters
•	Calculating reading time
•	Preview section of text
•	Converting text to lowercase
•	Auto Dismissing Alert messages
•	Enable DARK MODE feature
•	RESPONSIVE
TOOLS AND SOFTWARES USED:
As TextUtils is a REACT based text application. It requires some tools, software and frameworks in its developing process
•	HTML, CSS
•	REACT (JSX ALONG WITH ADVANCE JAVASCRIPT FEATURES)
•	VS Studio code
•	Node.js (which provides runtime javascript environment)